# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RagInt Backend API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/api/ask": {
      "post": {
        "summary": "Ask a question (SSE stream)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": { "type": "string" },
                  "request_id": { "type": "string" },
                  "client_id": { "type": "string" },
                  "kind": { "type": "string" },
                  "conversation_name": { "type": "string" },
                  "agent_id": { "type": "string" },
                  "guide": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/cancel": {
      "post": {
        "summary": "Cancel an active request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "request_id": { "type": "string" },
                  "client_id": { "type": "string" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
...
    "/api/speech_to_text": {
      "post": {
        "summary": "ASR (audio file upload)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio": { "type": "string", "format": "binary" },
                  "request_id": { "type": "string" },
                  "client_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
...
  }
}